requirements.txt
python-telegram-bot[http2]>=21.0.0
SQLAlchemy>=2.0
asyncpg>=0.29
